+++ src/main/kotlin/com/gemini/krakenbot/view/util/Icons.kt
@@ -20,7 +20,6 @@ object Icons {
     val SHIELD_EXCLAMATION = loadIcon("shield_exclamation.svg")
     val PLUS = loadIcon("plus.svg")
     val CHART = loadIcon("chart.svg")
-    val CLOCK = loadIcon("clock.svg")
 
     fun HTMLTag.icon(rawSvg: String) {
         unsafe { +rawSvg }